Implemented gradient calculation.
Starting to convert edge detection to use the gradient too, instead of
the simple 1d derivative. This is much better because the gradient
intensity will not depend on the edge direction.

The idea right now is to extract ~12 edgels, then use them for
orientation estimation along with the simple Kalman filter for blob
tracking, from before. Two totally separate filters. After that we
will see about integrating them into a single larger filter.